packages/projects/pyproject.toml
[project]
name = "rhoai-mcp-projects"
version = "0.1.0"
description = "Projects component for RHOAI MCP - Data Science Project management"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Adam Miller", email = "admiller@redhat.com" }
]
requires-python = ">=3.10"
keywords = ["mcp", "openshift", "rhoai", "projects", "namespace"]

dependencies = [
    "rhoai-mcp-core>=0.1.0,<1.0.0",
]

[project.optional-dependencies]
# Optional dependencies for full resource summary functionality
full = [
    "rhoai-mcp-notebooks>=0.1.0,<1.0.0",
    "rhoai-mcp-inference>=0.1.0,<1.0.0",
    "rhoai-mcp-pipelines>=0.1.0,<1.0.0",
]

[project.entry-points."rhoai_mcp.plugins"]
projects = "rhoai_mcp_projects.plugin:create_plugin"

[project.urls]
Homepage = "https://github.com/admiller/rhoai-mcp-prototype"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhoai_mcp_projects"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["rhoai_mcp_projects", "rhoai_mcp_core"]
